fix(stimme): alle 24 abgeschnittenen Voiceovers - Riegel an die echte Restlaenge
Nutzer-Auftrag 2026-09-14: "Außerdem gibt es diverse andere cutscenes wo die
voiceovers ebenfalls abgeschnitten werden, weil sie zu lang sind bevor das nächste
startet. Ermittele sie und verlängere dann jeweils den Bereich davor vor den Start
des nächsten Dialoges, damit der davor noch zu Ende gehen kann."

DIE ERMITTLUNG liegt vollstaendig unter analysis/voiceover_2026-09-14/ (Dossier +
Werkzeuge + Messsonde). Von 87 Aufnahmen in 9 Raeumen werden 24 abgeschnitten,
6 weitere sind knapp (unter 10 Bildern Reserve), 14 haben gar keinen Nachfolger und
laufen aus; unerreichbar ist keine einzige. Die zehn groessten Fehlbetraege:

  ROOM1150 main09   242,5 Bilder Aufnahme gegen 150 Skript-Zeit   fehlen 92,5
  ROOM10D0 main24   190,2                gegen 121                fehlen 69,2
  ROOM1150 main12   187,7                gegen 120                fehlen 67,7
  ROOM11B0 main08   176,2                gegen 110                fehlen 66,2
  ROOM10D0 main23   204,8                gegen 140                fehlen 64,8
  ROOM10D0 main18   149,1                gegen 101                fehlen 48,1
  ROOM1150 main13   176,2                gegen 130                fehlen 46,2
  ROOM11B0 main05   183,9                gegen 140                fehlen 43,9
  ROOM2000 main04   163,0                gegen 121                fehlen 42,0
  ROOM1150 main07   170,1                gegen 130                fehlen 40,1

Die Zeit zwischen zwei vertonten Zeilen besteht in ALLEN 24 Faellen ausschliesslich
aus Sleep-Ketten. Die Laengentabelle des Walkers stammt aus s_opcode_sizes des Ports,
die Blockgrenzen aus allen 24 Zeigern der RDT-Adresstabelle (ohne die laeuft ein
linearer Scan in den naechsten Block und erfindet Opcodes), und beides ist an
ROOM1170 sub02 gegengeprueft - die Folge msg0 / Sleep 50 / Sleep 51 / msg1 /
Sleep 100 kommt exakt heraus.

WAS SICH AM RIEGEL AENDERT - er verlaengert den Bereich davor, statt die Skriptdaten
anzufassen (die sind Original-Assets):

1. ER SITZT JETZT HINTER ALLEN ABFANGUNGEN, nicht mehr am Funktionsanfang. Vorher
   parkte er auch, was gar keine Dialogzeile ist: das SAVE-TELEFON (ROOM1150 msg 1)
   und die ITEM-BOX (msg 3) werden im Port abgefangen und springen mit pc+=4 heraus,
   der AUSWAHL-Dialog (ROOM1050 msg0, ROOM1090 msg8, ROOM10D0 msg7/11) parkt in
   seinem eigenen Zweig. Wer Telefon oder Box examinierte, waehrend noch eine
   Irons-Zeile klang, bekam bis zu 3 s lang gar keine Rueckmeldung.
2. DAS GATE AUF message_fsm_active FAELLT WEG. Es war zu eng: bei ROOM1150 main12
   steht der Text der Vorzeile beim naechsten Message_on noch (Standzeit 121 Bilder
   gegen 120 Bilder Abstand) - der Riegel griff dort gar nicht. An der neuen Stelle
   braucht es die Bedingung nicht, die parkenden Zweige liegen alle darueber.
3. GEWARTET WIRD GEGEN DIE ECHTE RESTLAENGE. Der Port kennt sie exakt:
   (s_xa.pcm_len - s_xa.pos) bei RE15_AUDIO_RATE, der Mischer rueckt pos ein Sample
   je Geraete-Frame vor. Die Plattform stempelt sie als g_re15_voice_restbilder im
   selben Block wie das Lauf-Flag. Der bisherige FESTE Deckel von 90 Bildern war mit
   der Begruendung "die laengste Aufnahme ist 117 Bilder" gesetzt - das galt fuer
   ROOM1170. Ueber den ganzen Bestand sind es 242,5 Bilder, und er haette ROOM1150
   main09 um 2,5 Bilder abgeschnitten. Geblieben ist eine Notbremse bei 300 Bildern
   gegen einen haengenden Kanal.
4. scd_thread_t.voice_wait ist jetzt 16 bit. Als uint8_t wrappte der Zaehler bei 255
   auf 0, und die 300er-Notbremse haette NIE ausgeloest - unit_voice_ausreden hat
   genau das gefunden.

Nebenbefund aus der Erhebung, ohne Wirkung: 68 der 87 WAV-Dateien tragen data-Groesse
0xFFFFFFFF (FFmpeg-Streaming-Artefakt, nicht nur room1170/main01). Die Klemme in
wav_find_data faengt das ab, und hinter data steht in diesen Dateien nichts mehr -
die Laengen stimmen also. Ein spaeterer Export mit nachlaufendem Chunk waere aber zu
lang; die Dateien einmal mit ffmpeg -c copy neu zu schreiben wuerde das erledigen.

Pin: tests/unit/test_voice_ausreden.c, jetzt neun Zusagen - darunter der
ROOM1150-main09-Fall mit 93 Bildern (den der alte Deckel abschnitt), die Notbremse,
und zwei neue: das Save-Telefon und der Auswahl-Dialog duerfen NICHT warten.
Dazu die Messsonde tests/unit/probe_stimme_takt.c aus der Erhebung.

Suite 295/295.

+++ re15_port/include/re15_scd.h
@@ -163,7 +163,10 @@ typedef struct {
      * Sleep-Werte im Skript bleiben unangetastet; gewartet wird nur, wenn eine
      * Aufnahme laenger ist als das Original ihr Zeit gibt - wo sie hineinpasst
      * (sechs von acht Zeilen), aendert sich NICHTS am Timing. */
-    uint8_t  voice_wait;
+    /* 16 bit, NICHT 8: die Notbremse des Riegels liegt bei 300 Bildern (ueber der
+     * laengsten Aufnahme des Bestands, 242,5). Als uint8_t wrappte der Zaehler bei 255
+     * auf 0 und die Notbremse haette nie ausgeloest - unit_voice_ausreden faengt das. */
+    uint16_t voice_wait;
 
     /* [#9] Switch (0x13) no longer snapshots a value into the thread: the
      * byte-true handler (LAB_8003fa5c) scans the whole Case/Default table inline
